backend/requirements.txt
fastapi
uvicorn
google-genai
qdrant-client
python-dotenv
asyncpg
psycopg2-binary
pydantic
pydantic-settings
python-jose[cryptography]
bcrypt
httpx
